feature: implement uDATMO node
Initial work for converting uDATMO from a library into a node with
event interface. The bulk of the code is written and compiles without
errors, but has not yet been tested, so in all likelihood it's buggy
as heck. Watch out for follow-up bugfix commits, as well as changes
to behavior_selector to utilize the new interface.

Signed-off-by: Helio Perroni Filho <mail@xperroni.me>