feature: Dual uDATMO interface
Previous work on udatmo2 focused on converting the module into a CARMEN
node. It has since been pointed that it might be preferable to keep the
option to use the module as a library, thus avoiding IPC costs. This
commits provides a dual interface allowing uDATMO to be used either as
a node or a library. This new interface will next be ported to udatmo1
and behavior_selector.

Signed-off-by: Helio Perroni Filho <mail@xperroni.me>