feature: Moving obstacle tracking
Previously uDATMO implemented moving obstacle detection but not
tracking -- it could not tell whether two obstacles found in successive
turns were the same. This commit extends the original model with the
concept of persistent moving obstacles, which are detected, tracked and
lost as they approach and distance themselves from the car. The ability
to connect successive observations to identified obstacles over time
will be necessary when obstacle detection is extended to encompass the
whole area around the car, not just the front.

Signed-off-by: Helio Perroni Filho <mail@xperroni.me>