bugfix: String returned by carmen_get_host() liable to change
The value of carmen_get_host() is retrieved through getenv(), which
returns a pointer to a static buffer whose contents may be changed by
later calls. Therefore it's not safe to keep (e.g. as a field value in
data structures). Fixed by creating a static variable to hold the host
name (which in theory may still be changed, but only if the host name
changes during the lifetime of the module).

See: http://www.cplusplus.com/reference/cstdlib/getenv/

"The pointer returned [by a getenv() call] points to an internal memory
block, whose content or validity may be altered by further calls to
getenv (but not by other library functions)."

Signed-off-by: Helio Perroni Filho <mail@xperroni.me>