feature: Function to get time of IPC initialization
When debugging CARMEN modules, it's often useful to know the time when
the program was started -- for example, so timestamps can be scaled
down to human-readable values. Function carmen_ipc_initialize_time()
returns the recorded time when carmen_ipc_initialize() was called. The
time is encoded as a timestamp, i.e. number of seconds since the Epoch.

Signed-off-by: Helio Perroni Filho <mail@xperroni.me>