feature: Improved moving obstacle handling
Module behavior_selector was updated so that a moving obstacle in front
of the vehicle is only taken into consideration for path planning if
it's closer than any other relevant elements. It's also able now to
perceive that a moving obstacle in front has stopped, and to react
accordingly (by stopping and waiting for it to move again). Lastly,
the moving obstacle simulator was overhauled and now runs much more
smoothly.