feature: Detect when moving obstacle shifts from side to front of car
Previously moving obstacles were reported without regard to their
positions relative to the car. This commit adds the ability to check
whether there is a moving obstacle in front of the car, ignoring moving
obstacles otherwise (moving obstacles to the sides of the car are still
tracked, so they can be seen in advance if they shift to its front).