feature: Bring udatmo2 up to speed with current state of source tree
Concerns about the performance of udatmo1 prompted a source rollback
that removed several DATMO API changes relied upon by udatmo2. This
commit adapts udatmo2 to the current state of the master branch, i.e.
using the older DATMO API's. It will be kept in a separate branch for
the time being, to be merged with the master branch at a later date.