feature: Remove reference-counted Track pointers
Initially it was thought that managing Track objects as individually
created heap instances was vital to avoid wasting processing time
swapping instances across collections. For that purpose a
reference-counted pointer type was defined and used everywhere a Track
object was required. Later it was noticed that most swap operations
could be avoided by using std::deque as default collection and swapping
to-be-deleted instances to the back of the collection then
pop_back()'ing them. Thus reference-counted Track pointers became an
unnecessary complication, and were therefore removed from the code.

Signed-off-by: Helio Perroni Filho <mail@xperroni.me>