bugfix: Scanning reading ranges that cross from quadrant 2 to 3
Scanning ranges of reading points that crossed from quadrant 2 to 3 was
originally impossible, due to the way points are indexed. Fixed by
adding a chained iterator type and Reading methods that use it to "glue"
together two point ranges, one from quadrant 2 to the border, the other
from the border into quadrant 3.

Signed-off-by: Helio Perroni Filho <mail@xperroni.me>